Fix phpstan level 0 error "unsafe usage of new static()"
Because a lot of these events / value objects use inheritance, which we cannot easily fix right now in scope of phpstan fixes, it was not possible to simply replace `static` with `self` and/or make the constructor of the abstract base class final. Instead it was sometimes necessary to remove the static method altogether and move it to the child classes.